+++ src/test/kotlin/org/jetbrains/intellij/tasks/SignPluginTaskSpec.kt
@@ -4,6 +4,7 @@ package org.jetbrains.intellij.tasks
 
 import org.jetbrains.intellij.IntelliJPluginConstants.DOWNLOAD_ZIP_SIGNER_TASK_NAME
 import org.jetbrains.intellij.IntelliJPluginConstants.SIGN_PLUGIN_TASK_NAME
+import org.jetbrains.intellij.IntelliJPluginConstants.VERIFY_PLUGIN_SIGNATURE_TASK_NAME
 import org.jetbrains.intellij.IntelliJPluginSpecBase
 import java.io.File
 import kotlin.test.Test
@@ -40,7 +41,7 @@ class SignPluginTaskSpec : IntelliJPluginSpecBase() {
     }
 
     @Test
-    fun `run Marketplace ZIP Signer in specified version`() {
+    fun `run Marketplace ZIP Signer in specified version using certificateChainFile and privateKeyFile`() {
         buildFile.groovy(
             """
             version = "1.0.0"
@@ -51,6 +52,90 @@ class SignPluginTaskSpec : IntelliJPluginSpecBase() {
             signPlugin {
                 certificateChainFile = file("${resolveResourcePath("certificates/cert.crt")}")
                 privateKeyFile = file("${resolveResourcePath("certificates/cert.key")}")
+            }
+            """.trimIndent()
+        )
+
+        build(SIGN_PLUGIN_TASK_NAME, "--info").let {
+            assertContains("marketplace-zip-signer-cli.jar", it.output)
+        }
+    }
+
+    @Test
+    fun `fail on signing without password provided`() {
+        buildFile.groovy(
+            """
+            version = "1.0.0"
+            
+            downloadZipSigner {
+                version = "0.1.7"
+            }
+            signPlugin {
+                certificateChainFile = file("${resolveResourcePath("certificates-password/chain.crt")}")
+                privateKeyFile = file("${resolveResourcePath("certificates-password/private_encrypted.pem")}")
+            }
+            """.trimIndent()
+        )
+
+        buildAndFail(SIGN_PLUGIN_TASK_NAME).let {
+            assertContains("Can't read private key. Password is missing", it.output)
+        }
+    }
+
+    @Test
+    fun `fail on signing with incorrect password provided`() {
+        buildFile.groovy(
+            """
+            version = "1.0.0"
+            
+            downloadZipSigner {
+                version = "0.1.7"
+            }
+            signPlugin {
+                certificateChainFile = file("${resolveResourcePath("certificates-password/chain.crt")}")
+                privateKeyFile = file("${resolveResourcePath("certificates-password/private_encrypted.pem")}")
+                password = "incorrect"
+            }
+            """.trimIndent()
+        )
+
+        buildAndFail(SIGN_PLUGIN_TASK_NAME).let {
+            assertContains("unable to read encrypted data", it.output)
+        }
+    }
+
+    @Test
+    fun `sign plugin with password provided`() {
+        buildFile.groovy(
+            """
+            version = "1.0.0"
+            
+            downloadZipSigner {
+                version = "0.1.7"
+            }
+            signPlugin {
+                certificateChainFile = file("${resolveResourcePath("certificates-password/chain.crt")}")
+                privateKeyFile = file("${resolveResourcePath("certificates-password/private_encrypted.pem")}")
+                password = "foobar"
+            }
+            """.trimIndent()
+        )
+
+        build(SIGN_PLUGIN_TASK_NAME)
+    }
+
+    @Test
+    fun `run Marketplace ZIP Signer in specified version using certificateChain and privateKey`() {
+        buildFile.groovy(
+            """
+            version = "1.0.0"
+            
+            downloadZipSigner {
+                version = "0.1.7"
+            }
+            signPlugin {
+                certificateChain = ${"\"\"\""}${resolveResourceContent("certificates/cert.crt")}${"\"\"\""}
+                privateKey = ${"\"\"\""}${resolveResourceContent("certificates/cert.key")}${"\"\"\""}
             }
             """.trimIndent()
         )
@@ -113,5 +198,98 @@ class SignPluginTaskSpec : IntelliJPluginSpecBase() {
         assertTrue(distributionFolder.listFiles()?.asList()?.any {
             it.name.equals("projectName-1.0.0-signed.zip")
         } ?: false)
+    }
+
+    @Test
+    fun `skip plugin signature verification task if plugin signing is not configured`() {
+        buildFile.groovy(
+            """
+            version = "1.0.0"
+            """.trimIndent()
+        )
+
+        build(VERIFY_PLUGIN_SIGNATURE_TASK_NAME).let {
+            assertContains("Task :$SIGN_PLUGIN_TASK_NAME SKIPPED", it.output)
+            assertContains("Task :$VERIFY_PLUGIN_SIGNATURE_TASK_NAME NO-SOURCE", it.output)
+        }
+    }
+
+    @Test
+    fun `verify plugin signature with certificateChainFile`() {
+        buildFile.groovy(
+            """
+            version = "1.0.0"
+            
+            signPlugin {
+                certificateChainFile = file("${resolveResourcePath("certificates/cert.crt")}")
+                privateKeyFile = file("${resolveResourcePath("certificates/cert.key")}")
+            }
+            """.trimIndent()
+        )
+        build(VERIFY_PLUGIN_SIGNATURE_TASK_NAME)
+    }
+
+    @Test
+    fun `verify plugin signature with certificateChain`() {
+        buildFile.groovy(
+            """
+            version = "1.0.0"
+            
+            downloadZipSigner {
+                version = "0.1.7"
+            }
+            signPlugin {
+                certificateChain = ${"\"\"\""}${resolveResourceContent("certificates/cert.crt")}${"\"\"\""}
+                privateKey = ${"\"\"\""}${resolveResourceContent("certificates/cert.key")}${"\"\"\""}
+            }
+            """.trimIndent()
+        )
+
+        build(VERIFY_PLUGIN_SIGNATURE_TASK_NAME)
+    }
+
+    @Test
+    fun `verify plugin signed with password provided`() {
+        buildFile.groovy(
+            """
+            version = "1.0.0"
+            
+            downloadZipSigner {
+                version = "0.1.7"
+            }
+            signPlugin {
+                certificateChainFile = file("${resolveResourcePath("certificates-password/chain.crt")}")
+                privateKeyFile = file("${resolveResourcePath("certificates-password/private_encrypted.pem")}")
+                password = "foobar"
+            }
+            """.trimIndent()
+        )
+
+        build(VERIFY_PLUGIN_SIGNATURE_TASK_NAME)
+    }
+
+    @Test
+    fun `verify unsigned plugin archive`() {
+        buildFile.groovy(
+            """
+            version = "1.0.0"
+            
+            downloadZipSigner {
+                version = "0.1.7"
+            }
+            signPlugin {
+                certificateChainFile = file("${resolveResourcePath("certificates-password/chain.crt")}")
+                privateKeyFile = file("${resolveResourcePath("certificates-password/private_encrypted.pem")}")
+                password = "foobar"
+            }
+            verifyPluginSignature {
+                inputArchiveFile = buildPlugin.archiveFile
+            }
+            """.trimIndent()
+        )
+
+        buildAndFail(VERIFY_PLUGIN_SIGNATURE_TASK_NAME).let {
+            assertContains("Provided zip archive is not signed", it.output)
+        }
     }
 }